Refactor Navidrome client and md5
