+++ static/js/navidrome/constants.js
@@ -1,0 +1,6 @@
+export const NAVIDROME_API_VERSION = '1.16.1';
+export const NAVIDROME_CLIENT_ID = 'listenbrainz-wrapped';
+export const NAVIDROME_ALBUM_PAGE_SIZE = 500;
+export const NAVIDROME_ALBUM_DELAY_MS = 150;
+export const NAVIDROME_DETAIL_DELAY_MS = 80;
+export const NAVIDROME_MAX_ALBUMS_TO_SCAN = 2000;
